+++ scripts/dev-bundle-tool-package.js
@@ -14,14 +14,15 @@ var packageJson = {
     npm: "4.4.4",
     "node-gyp": "3.6.0",
     "node-pre-gyp": "0.6.34",
-    "meteor-babel": "0.20.1",
+    "meteor-babel": "0.20.2",
+    reify: "0.8.2",
     "meteor-promise": "0.8.0",
     fibers: "1.0.15",
     promise: "7.1.1",
     // So that Babel 6 can emit require("babel-runtime/helpers/...") calls.
     "babel-runtime": "6.9.2",
     // For various ES2015 polyfills, such as Map and Set.
-    "meteor-ecmascript-runtime": "0.2.9",
+    "meteor-ecmascript-runtime": "0.3.0",
     // Not yet upgrading Underscore from 1.5.2 to 1.7.0 (which should be done
     // in the package too) because we should consider using lodash instead
     // (and there are backwards-incompatible changes either way).